pyproject.toml
[project]
name = "chuqin"
version = "0.0.1"
readme = "README.md"

requires-python = ">=3.11"

dependencies = [
    "pywebview>=5.0.0",
]

[project.optional-dependencies]
build = [
    "pyinstaller>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/hezhangjian/ChuQin-PyWebView"
Repository = "https://github.com/hezhangjian/ChuQin-PyWebView"
Documentation = "https://github.com/hezhangjian/ChuQin-PyWebView#readme"

[tool.mypy]
python_version = "3.11"

[tool.ruff]
line-length = 120
